open server socket and request handle scaffold

+++ crates/bevy_quinnet/src/shared.rs
@@ -1,15 +1,16 @@
 use std::{fmt, io, net::AddrParseError, sync::PoisonError};
 
-use crate::client::ConnectionId;
 use bevy::prelude::{Deref, DerefMut, Resource};
 use rcgen::RcgenError;
 use tokio::runtime::Runtime;
+
+use crate::client::ConnectionId;
 
 pub const DEFAULT_MESSAGE_QUEUE_SIZE: usize = 150;
 pub const DEFAULT_KILL_MESSAGE_QUEUE_SIZE: usize = 10;
 pub const DEFAULT_KEEP_ALIVE_INTERVAL_S: u64 = 4;
 
-pub type ClientId = u64;
+pub type ClientId = u32;
 
 #[derive(Resource, Deref, DerefMut)]
 pub struct AsyncRuntime(pub Runtime);
